Mon, Nov 14, 2022, 8:01 PM +00:00

+++ chapter2/readability/readability.c
@@ -31,8 +31,8 @@ int main(void)
     float averagew = (float)sentences / words * 100;
 
     // Rounding
-    round(averagel);
-    round(averagew);
+    roundf(averagel);
+    roundf(averagew);
 
     // TODO: Calculate the Coleman-Liau index
     int level = (0.0588 * average - 0.296 * sentences - 15.8);
